refactor: Move Foods from feature_review to food package
